pkg/warp: move a lot of unnecessary packages inside.
